requirements.txt
fastapi
uvicorn
openai
graphviz
pydantic
python-dotenv